JAMES-2688 Move FakePropertiesProvider to guice-common
