+++ server/container/guice/guice-common/src/test/java/org/apache/james/FakePropertiesProvider.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.modules.objectstorage;
+package org.apache.james;
 
 import java.io.FileNotFoundException;
